requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==11.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
dateparser==1.2.0

# AI/LLM for text extraction (optional)
openai==1.6.1

# Rate limiting and caching
ratelimit==2.2.1
cachetools==5.3.2

# Export formats
openpyxl==3.1.2
jinja2==3.1.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: OpenCorporates support
# requests-cache==1.1.1
